Moved metrics setups to separate file
